Added market functions. Tested market item addition.

+++ ItemContract/Structs.cs
@@ -79,7 +79,7 @@ namespace LordsContract
     {
         public BigInteger ID;                   // Stronghold ID
         public BigInteger Hero;                 // Hero ID, that occupies City on map
-        public decimal Coffer;               // City Coffer
+        public BigInteger Coffer;               // City Coffer
         public BigInteger Size;
         public BigInteger ItemsOnMarket;        // Current amount of items on city market
         public BigInteger ItemsCap;
